backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
mediapipe>=0.10.14
opencv-python-headless>=4.9.0
numpy>=1.26.0
Pillow>=10.2.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
groq>=0.9.0
ultralytics>=8.0.0
